git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/trunk@69619 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/color/colortraits.h
@@ -185,7 +185,7 @@ namespace anatomist
   AimsRGBA ColorScalarPaletteTraits<T>::color( const T & in ) const
   {
     int ival0, ival1;
-    float fval0, fval1, val = static_cast<float>( in );
+    double fval0, fval1, val = static_cast<double>( in );
 
     // Comparisons are written this way to accommodate NaN and Inf
     fval0 = scale0 * val + decal0;
